chore(router): inline backend_args.py into args.py, one less mystery file
